Move WordPress admin functionality to its own class
